+++ package.json
@@ -1,7 +1,6 @@
 {
   "name": "web3dash_2",
   "version": "0.1.0",
-  "homepage": "https://lostconversation.github.io/web3dash2",
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^5.16.5",
